// angular/src/app/application/parts-management/status-color.enum.ts
export enum StatusColor {
  PREPICK = '#FFC000',
  PREPICK_OLD = '#FF00C0',
  CANCEL_PREPICK = '#FF8000',
  CANCEL_PREPICK_OLD = '#FF00C0',
  COMPLETE = '#00C000',
  RETURN = '#FF0000',
}

export enum StatusColorLabel {
  PREPICK = 'STATUS_PREPICK',
  PREPICK_OLD = 'STATUS_PREPICK_OLD',
  CANCEL_PREPICK = 'STATUS_CANCEL_PREPICK',
  CANCEL_PREPICK_OLD = 'STATUS_CANCEL_PREPICK_OLD',
  COMPLETE = 'STATUS_COMPLETE',
  RETURN = 'STATUS_RETURN',
}